chore: go pipeline fix

+++ src/main/java/com/twilio/oai/TwilioGoGenerator.java
@@ -190,14 +190,14 @@ public class TwilioGoGenerator extends AbstractTwilioGoGenerator {
                 // make sure the format matches the other methods
                 co.vendorExtensions.put("x-domain-name", co.nickname.replaceFirst("List", ""));
                 co.allParams.forEach( p -> {
-                    if( p.dataType.equals("int") && ( p.baseName.equals("PageSize") || p.baseName.equals("limit") ) ){
-                        p.dataType = "int64";
+                    if( p.dataType.equals("int64") && ( p.baseName.equals("PageSize") || p.baseName.equals("limit") ) ){
+                        p.dataType = "int";
                     }
                 });
 
                 co.optionalParams.forEach( p -> {
-                    if( p.dataType.equals("int") && ( p.baseName.equals("PageSize") || p.baseName.equals("limit") ) ){
-                        p.dataType = "int64";
+                    if( p.dataType.equals("int64") && ( p.baseName.equals("PageSize") || p.baseName.equals("limit") ) ){
+                        p.dataType = "int";
                     }
                 });
